requirements.txt
vscode-ext==1.4.0
youtube-search-python